# pkg/server/api/docs/swagger.yaml
components:
  schemas:
...
    config.AnycastConfig:
      properties:
        services:
          items:
            $ref: '#/components/schemas/config.AnycastService'
          type: array
          uniqueItems: false
      type: object
...
    config.AnycastService:
      properties:
        active:
          type: boolean
        anycast_ips:
          items:
            type: string
          type: array
          uniqueItems: false
        endpoints:
          items:
            $ref: '#/components/schemas/config.AnycastEndpoint'
          type: array
          uniqueItems: false
        name:
          type: string
      required:
      - name
      type: object
    config.BFD:
      properties:
        profiles:
          items:
            $ref: '#/components/schemas/config.BFDProfile'
          type: array
          uniqueItems: false
      type: object
    config.BFDProfile:
      properties:
        detect_multiplier:
          type: integer
        echo_mode:
          type: boolean
        extra:
          items:
            type: string
          type: array
          uniqueItems: false
        minimum_ttl:
          type: integer
        name:
          type: string
        passive_mode:
          type: boolean
        receive_interval:
          type: integer
        transmit_interval:
          type: integer
      required:
      - name
      type: object
    config.BGP:
      properties:
        address_families:
          additionalProperties:
            $ref: '#/components/schemas/config.BGPAddressFamily'
          type: object
        allow_inbound:
          items:
            type: string
          type: array
          uniqueItems: false
        asn:
          type: integer
        extra:
          items:
            type: string
          type: array
          uniqueItems: false
        neighbors:
          items:
            $ref: '#/components/schemas/config.BGPNeighbor'
          type: array
          uniqueItems: false
        no_default_ipv4_unicast:
          type: boolean
        no_ebgp_requires_policy:
          type: boolean
        no_import_check:
          type: boolean
        no_rib:
          type: boolean
        prefix_lists:
          additionalProperties:
            items:
              $ref: '#/components/schemas/config.PrefixEntry'
            type: array
          type: object
        route_maps:
          additionalProperties:
            items:
              $ref: '#/components/schemas/config.RouteMapEntry'
            type: array
          type: object
        router_id:
          type: string
      required:
      - asn
      type: object
...
    config.Bond:
      properties:
        downdelay:
          type: integer
        lacp_rate:
          type: string
        members:
          items:
            type: string
          type: array
          uniqueItems: false
        miimon:
          type: integer
        min_links:
          type: integer
        mode:
          type: string
        primary:
          type: string
        updelay:
          type: integer
        xmit_hash_policy:
          type: string
      type: object
    config.Bridge:
      properties:
        members:
          items:
            type: string
          type: array
          uniqueItems: false
        stp:
          type: boolean
      type: object
    config.CollectionIntervals:
      properties:
        bgp:
          type: integer
        iface:
          type: integer
        lldp:
          type: integer
        neighbors:
          type: integer
        proto:
          type: integer
        routes:
          type: integer
        system:
          type: integer
      type: object
    config.Config:
      properties:
        boot_modules:
          items:
            type: string
          type: array
          uniqueItems: false
        dhcp:
          $ref: '#/components/schemas/config.DHCP'
        dns:
          $ref: '#/components/schemas/config.DNS'
        friends:
          items:
            $ref: '#/components/schemas/config.Friend'
          type: array
          uniqueItems: false
        gai:
          $ref: '#/components/schemas/config.GAIConfig'
        ha:
          $ref: '#/components/schemas/config.HA'
        hostname:
          type: string
        interfaces:
          additionalProperties:
            $ref: '#/components/schemas/config.Interface'
          type: object
        logging:
          $ref: '#/components/schemas/config.Logging'
        monitoring:
          $ref: '#/components/schemas/config.MonitoringConfig'
        nftables:
          $ref: '#/components/schemas/config.NftablesConfig'
        routing:
          $ref: '#/components/schemas/config.Routing'
        services:
          additionalProperties:
            $ref: '#/components/schemas/config.Service'
          type: object
        ssh:
          $ref: '#/components/schemas/config.SSH'
        sysctl:
          additionalProperties:
            type: string
          type: object
        tunnels:
          additionalProperties:
            $ref: '#/components/schemas/config.Tunnel'
          type: object
        users:
          additionalProperties:
            $ref: '#/components/schemas/config.User'
          type: object
        version:
          type: string
        vrfs:
          additionalProperties:
            $ref: '#/components/schemas/config.VRFConfig'
          type: object
        wireguard:
          additionalProperties:
            $ref: '#/components/schemas/config.Wireguard'
          type: object
      required:
      - hostname
      - version
      type: object
    config.Conntrackd:
      properties:
        address:
          type: string
        allow_inbound:
          type: boolean
        interface:
          type: string
        peer_ips:
          items:
            type: string
          type: array
          uniqueItems: false
        port:
          type: integer
      required:
      - address
      - interface
      type: object
    config.DHCP:
      properties:
        control_agent:
          $ref: '#/components/schemas/config.KeaControlAgent'
        ddns:
          $ref: '#/components/schemas/config.DHCPDDNS'
        enabled:
          type: boolean
        interfaces:
          items:
            type: string
          type: array
          uniqueItems: false
        subnets4:
          items:
            $ref: '#/components/schemas/config.KeaSubnet'
          type: array
          uniqueItems: false
        subnets6:
          items:
            $ref: '#/components/schemas/config.KeaSubnet'
          type: array
          uniqueItems: false
      type: object
    config.DHCPDDNS:
      properties:
        algorithm:
          type: string
        domain:
          type: string
        enabled:
          type: boolean
        generated_prefix:
          type: string
        hostname_char_set:
          type: string
        key:
          type: string
        override_client_update:
          type: boolean
        override_no_update:
          type: boolean
        replace_client_name:
          type: string
        reverse:
          type: boolean
        ttl:
          type: integer
      type: object
    config.DNS:
      properties:
        nameservers:
          items:
            type: string
          type: array
          uniqueItems: false
        search:
          items:
            type: string
          type: array
          uniqueItems: false
        server:
          $ref: '#/components/schemas/config.DNSServer'
      type: object
    config.DNSCache:
      properties:
        disabled:
          type: boolean
        max_negative_ttl:
          type: integer
        max_ttl:
          type: integer
        min_ttl:
          type: integer
        prefetch:
          type: boolean
        rrset_size:
          type: string
        serve_expired:
          type: boolean
        size:
          type: string
      type: object
    config.DNSForward:
      properties:
        dnssec:
          type: boolean
        domain:
          type: string
        first_try:
          type: boolean
        servers:
          items:
            type: string
          type: array
          uniqueItems: false
        tls_upstream:
          type: boolean
      required:
      - domain
      type: object
...
    config.DNSServer:
      properties:
        allow_from:
          items:
            type: string
          type: array
          uniqueItems: false
        allow_inbound:
          items:
            type: string
          type: array
          uniqueItems: false
        cache:
          $ref: '#/components/schemas/config.DNSCache'
        dnssec:
          type: boolean
        enabled:
          type: boolean
        extra:
          items:
            type: string
          type: array
          uniqueItems: false
        forward:
          items:
            $ref: '#/components/schemas/config.DNSForward'
          type: array
          uniqueItems: false
        listen:
          items:
            type: string
          type: array
          uniqueItems: false
        log_queries:
          type: boolean
        mode:
          type: string
        port:
          type: integer
        threads:
          type: integer
        upstreams:
          items:
            type: string
          type: array
          uniqueItems: false
        views:
          items:
            $ref: '#/components/schemas/config.DNSView'
          type: array
          uniqueItems: false
        zones:
          items:
            $ref: '#/components/schemas/config.DNSZone'
          type: array
          uniqueItems: false
      type: object
    config.DNSView:
      properties:
        extra:
          items:
            type: string
          type: array
          uniqueItems: false
        forward:
          items:
            $ref: '#/components/schemas/config.DNSForward'
          type: array
          uniqueItems: false
        match_from:
          items:
            type: string
          type: array
          uniqueItems: false
        name:
          type: string
        recursion:
          type: boolean
        upstreams:
          items:
            type: string
          type: array
          uniqueItems: false
        zones:
          items:
            $ref: '#/components/schemas/config.DNSZone'
          type: array
          uniqueItems: false
      required:
      - name
      type: object
    config.DNSZone:
      properties:
        dnssec:
          type: boolean
        extra:
          items:
            type: string
          type: array
          uniqueItems: false
        for_upstream:
          type: boolean
        name:
          type: string
        nameservers:
          items:
            type: string
          type: array
          uniqueItems: false
        primaries:
          items:
            type: string
          type: array
          uniqueItems: false
        records:
          items:
            $ref: '#/components/schemas/config.DNSRecord'
          type: array
          uniqueItems: false
        soa:
          $ref: '#/components/schemas/config.DNSSOA'
        ttl:
          type: integer
      required:
      - name
      type: object
    config.Friend:
      properties:
        enabled:
          type: boolean
        ha:
          $ref: '#/components/schemas/config.FriendHA'
        hostname:
          type: string
        identity:
          $ref: '#/components/schemas/config.FriendIdentity'
        manage:
          type: boolean
        name:
          type: string
        sync:
          $ref: '#/components/schemas/config.FriendSync'
        tls_skip_verify:
          type: boolean
        token:
          type: string
        url:
          type: string
      required:
      - name
      - token
      - url
      type: object
    config.FriendHA:
      properties:
        enabled:
          type: boolean
        link:
          $ref: '#/components/schemas/config.FriendHALink'
        priority:
          type: integer
      type: object
    config.FriendHALink:
      properties:
        address:
          type: string
        interface:
          type: string
        port:
          type: integer
      required:
      - address
      - interface
      type: object
    config.FriendIdentity:
      properties:
        fingerprint:
          type: string
        public_key:
          type: string
        x25519_public_key:
          type: string
      type: object
    config.FriendSync:
      properties:
        overrides:
          additionalProperties:
            type: string
          type: object
        sections:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    config.GAIConfig:
      properties:
        label:
          items:
            $ref: '#/components/schemas/config.GAIEntry'
          type: array
          uniqueItems: false
        precedence:
          items:
            $ref: '#/components/schemas/config.GAIEntry'
          type: array
          uniqueItems: false
      type: object
    config.GAIEntry:
      properties:
        prefix:
          type: string
        value:
          type: integer
      required:
      - prefix
      - value
      type: object
    config.HA:
      properties:
        conntrackd:
          $ref: '#/components/schemas/config.Conntrackd'
        vrrp:
          items:
            $ref: '#/components/schemas/config.VRRPInstance'
          type: array
          uniqueItems: false
      type: object
    config.Interface:
      properties:
        addresses:
          items:
            type: string
          type: array
          uniqueItems: false
        bond:
          $ref: '#/components/schemas/config.Bond'
        bridge:
          $ref: '#/components/schemas/config.Bridge'
        dhcp_options:
          items:
            type: string
          type: array
          uniqueItems: false
        mtu:
          type: integer
        select:
          type: string
        type:
          type: string
        vlan:
          $ref: '#/components/schemas/config.VLAN'
        vrf:
          type: string
        vxlan:
          $ref: '#/components/schemas/config.VXLAN'
      type: object
    config.KeaControlAgent:
      properties:
        password:
          type: string
        url:
          type: string
        user:
          type: string
      type: object
    config.KeaReservation:
      properties:
        duid:
          type: string
        hostname:
          type: string
        hw_address:
          type: string
        ip_address:
          type: string
      type: object
    config.KeaSubnet:
      properties:
        ddns:
          type: boolean
        ddns_domain:
          type: string
        dns:
          items:
            type: string
          type: array
          uniqueItems: false
        exclusions:
          items:
            type: string
          type: array
          uniqueItems: false
        gateway:
          type: string
        interface:
          type: string
        pools:
          items:
            type: string
          type: array
          uniqueItems: false
        reservations:
          items:
            $ref: '#/components/schemas/config.KeaReservation'
          type: array
          uniqueItems: false
        subnet:
          type: string
        valid_lifetime:
          type: integer
      required:
      - subnet
      type: object
    config.LLDPConfig:
      properties:
        cdp:
          type: boolean
        enabled:
          type: boolean
        interfaces:
          items:
            type: string
          type: array
          uniqueItems: false
        transmit:
          type: boolean
      type: object
    config.Logging:
      properties:
        file:
          type: string
        level:
          type: string
        target:
          type: string
      type: object
    config.ManagedRule:
      properties:
        action:
          type: string
        action_to:
          type: string
        comment:
          type: string
        disabled:
          type: boolean
        match:
          $ref: '#/components/schemas/config.RuleMatch'
        tag:
          type: string
      required:
      - action
      type: object
    config.MonitoringConfig:
      properties:
        collection:
          $ref: '#/components/schemas/config.CollectionIntervals'
        lldp:
          $ref: '#/components/schemas/config.LLDPConfig'
        probes:
          items:
            $ref: '#/components/schemas/config.PingProbe'
          type: array
          uniqueItems: false
      type: object
    config.NftChain:
      properties:
        files:
          items:
            type: string
          type: array
          uniqueItems: false
        managed:
          items:
            $ref: '#/components/schemas/config.ManagedRule'
          type: array
          uniqueItems: false
        policy:
          type: string
        rules:
          type: string
      type: object
    config.NftablesConfig:
      properties:
        chains:
          additionalProperties:
            $ref: '#/components/schemas/config.NftChain'
          type: object
        defines:
          type: string
        include:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    config.OSPF:
      properties:
        allow_inbound:
          items:
            type: string
          type: array
          uniqueItems: false
        areas:
          items:
            $ref: '#/components/schemas/config.OSPFArea'
          type: array
          uniqueItems: false
        default_information_originate:
          type: boolean
        distance:
          type: integer
        extra:
          items:
            type: string
          type: array
          uniqueItems: false
        interfaces:
          additionalProperties:
            $ref: '#/components/schemas/config.OSPFInterface'
          type: object
        passive_interfaces:
          items:
            type: string
          type: array
          uniqueItems: false
        redistribute:
          items:
            type: string
          type: array
          uniqueItems: false
        reference_bandwidth:
          type: integer
        router_id:
          type: string
      type: object
    config.OSPF6:
      properties:
        allow_inbound:
          items:
            type: string
          type: array
          uniqueItems: false
        areas:
          items:
            $ref: '#/components/schemas/config.OSPF6Area'
          type: array
          uniqueItems: false
        default_information_originate:
          type: boolean
        distance:
          type: integer
        extra:
          items:
            type: string
          type: array
          uniqueItems: false
        interfaces:
          additionalProperties:
            $ref: '#/components/schemas/config.OSPF6Interface'
          type: object
        passive_interfaces:
          items:
            type: string
          type: array
          uniqueItems: false
        redistribute:
          items:
            type: string
          type: array
          uniqueItems: false
        reference_bandwidth:
          type: integer
        router_id:
          type: string
      type: object
...
    config.PBR:
      properties:
        maps:
          additionalProperties:
            items:
              $ref: '#/components/schemas/config.PBRMapEntry'
            type: array
          type: object
        nexthop_groups:
          additionalProperties:
            $ref: '#/components/schemas/config.PBRNexthopGroup'
          type: object
        policies:
          additionalProperties:
            type: string
          type: object
      type: object
...
    config.PingProbe:
      properties:
        interval:
          type: integer
        name:
          type: string
        target:
          type: string
        timeout:
          type: integer
      required:
      - name
      - target
      type: object
    config.PrefixEntry:
      properties:
        action:
          type: string
        ge:
          type: integer
        le:
          type: integer
        prefix:
          type: string
        seq:
          type: integer
      required:
      - action
      - prefix
      - seq
      type: object
    config.RADVDConfig:
      properties:
        interfaces:
          additionalProperties:
            $ref: '#/components/schemas/config.RADVDInterface'
          type: object
      type: object
...
    config.RouteMapEntry:
      properties:
        action:
          type: string
        call:
          type: string
        continue:
          type: integer
        match:
          additionalProperties:
            type: string
          type: object
        on_match:
          type: string
        seq:
          type: integer
        set:
          additionalProperties:
            type: string
          type: object
      required:
      - action
      - seq
      type: object
    config.Routing:
      properties:
        anycast:
          $ref: '#/components/schemas/config.AnycastConfig'
        bfd:
          $ref: '#/components/schemas/config.BFD'
        bgp:
          $ref: '#/components/schemas/config.BGP'
        ospf:
          $ref: '#/components/schemas/config.OSPF'
        ospf6:
          $ref: '#/components/schemas/config.OSPF6'
        pbr:
          $ref: '#/components/schemas/config.PBR'
        radvd:
          $ref: '#/components/schemas/config.RADVDConfig'
        static:
          items:
            $ref: '#/components/schemas/config.StaticRoute'
          type: array
          uniqueItems: false
        vrfs:
          additionalProperties:
            $ref: '#/components/schemas/config.VRFRouting'
          type: object
      type: object
...
    config.SSH:
      properties:
        allow_groups:
          items:
            type: string
          type: array
          uniqueItems: false
        allow_tcp_forwarding:
          type: string
        allow_users:
          items:
            type: string
          type: array
          uniqueItems: false
        banner:
          type: string
        client_alive_count_max:
          type: integer
        client_alive_interval:
          type: integer
        login_grace_time:
          type: integer
        max_auth_tries:
          type: integer
        password_auth:
          type: string
        permit_root_login:
          type: string
        port:
          type: integer
        pubkey_auth:
          type: string
        x11_forwarding:
          type: string
      type: object
    config.Service:
      properties:
        after:
          type: string
        configs:
          items:
            $ref: '#/components/schemas/config.ServiceConfig'
          type: array
          uniqueItems: false
        enable:
          type: boolean
      required:
      - enable
      type: object
    config.ServiceConfig:
      properties:
        dest:
          type: string
        mode:
          type: integer
        src:
          type: string
      required:
      - dest
      - src
      type: object
    config.StaticRoute:
      properties:
        destination:
          type: string
        dev:
          type: string
        metric:
          type: integer
        via:
          type: string
      required:
      - destination
      type: object
    config.Tunnel:
      properties:
        addresses:
          items:
            type: string
          type: array
          uniqueItems: false
        friend:
          type: string
        local:
          type: string
        mode:
          type: string
        mtu:
          type: integer
        remote:
          type: string
        ttl:
          type: integer
      required:
      - mode
      type: object
    config.User:
      properties:
        groups:
          items:
            type: string
          type: array
          uniqueItems: false
        home:
          type: string
        password_hash:
          type: string
        shell:
          type: string
        ssh_keys:
          items:
            type: string
          type: array
          uniqueItems: false
        system:
          type: boolean
        uid:
          type: integer
      type: object
    config.VLAN:
      properties:
        id:
          type: integer
      required:
      - id
      type: object
    config.VRFConfig:
      properties:
        table:
          type: integer
      required:
      - table
      type: object
    config.VRFRouting:
      properties:
        bgp:
          $ref: '#/components/schemas/config.BGP'
        ospf:
          $ref: '#/components/schemas/config.OSPF'
        ospf6:
          $ref: '#/components/schemas/config.OSPF6'
        static:
          items:
            $ref: '#/components/schemas/config.StaticRoute'
          type: array
          uniqueItems: false
      type: object
    config.VRRPInstance:
      properties:
        allow_inbound:
          type: boolean
        friend:
          type: string
        id:
          type: integer
        interface:
          type: string
        name:
          type: string
        password:
          type: string
        priority:
          type: integer
        switchover:
          type: boolean
        track_interfaces:
          items:
            type: string
          type: array
          uniqueItems: false
        transport:
          type: string
        vips:
          items:
            type: string
          type: array
          uniqueItems: false
        virtual_routes:
          items:
            type: string
          type: array
          uniqueItems: false
      required:
      - id
      - interface
      type: object
    config.VXLAN:
      properties:
        external:
          type: boolean
        group:
          type: string
        learning:
          type: boolean
        local:
          type: string
        port:
          type: integer
        remote:
          type: string
        vni:
          type: integer
        vnifilter:
          type: boolean
        vtep:
          type: string
      type: object
    config.WGPeer:
      properties:
        allowed_ips:
          items:
            type: string
          type: array
          uniqueItems: false
        endpoint:
          type: string
        keepalive:
          type: integer
        name:
          type: string
        preshared_key:
          type: string
        preshared_key_file:
          type: string
        private_key:
          type: string
        public_key:
          type: string
      type: object
    config.Wireguard:
      properties:
        addresses:
          items:
            type: string
          type: array
          uniqueItems: false
        allow_inbound:
          type: boolean
        friend:
          type: string
        listen_port:
          type: integer
        mtu:
          type: integer
        peers:
          items:
            $ref: '#/components/schemas/config.WGPeer'
          type: array
          uniqueItems: false
        post_down:
          items:
            type: string
          type: array
          uniqueItems: false
        post_up:
          items:
            type: string
          type: array
          uniqueItems: false
        pre_down:
          items:
            type: string
          type: array
          uniqueItems: false
        pre_up:
          items:
            type: string
          type: array
          uniqueItems: false
        private_key:
          type: string
        private_key_file:
          type: string
        table:
          type: string
      type: object
...
    types.Response-config_Config:
      properties:
        code:
          type: integer
        error:
          type: string
        result:
          $ref: '#/components/schemas/config.Config'
        status:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: '"Bearer <token>"'
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: Configuration and monitoring API for a Routier node.
  title: Routier API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/config:
    get:
      parameters:
      - description: configuration layer
        in: query
        name: layer
        schema:
          enum:
          - advanced
          - simple
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/types.Response-config_Config'
          description: OK
      security:
      - BearerAuth: []
      summary: Get the staged config
      tags:
      - config
...
  /api/system/updates:
    get:
      parameters:
      - description: force a fresh apk index refresh
        in: query
        name: refresh
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: OK
      security:
      - BearerAuth: []
      summary: List available package upgrades
      tags:
      - system
...
  /api/v1/config:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: OK
      security:
      - BearerAuth: []
      summary: Get the committed canonical configuration as YAML
      tags:
      - v1-config
...
  /api/v1/sessions/{sessionID}/config:
    get:
      parameters:
      - description: session id
        in: path
        name: sessionID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: OK
      security:
      - BearerAuth: []
      summary: Get a session's canonical configuration as YAML
      tags:
      - v1-config
    put:
      parameters:
      - description: session id
        in: path
        name: sessionID
        required: true
        schema:
          type: string
      requestBody:
        content:
          text/plain:
            schema:
              type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
          description: OK
      security:
      - BearerAuth: []
      summary: Replace a session's canonical configuration with YAML
      tags:
      - v1-config
...
servers:
- url: /